+++ src/main.ts
@@ -92,5 +92,4 @@ program
     }
   });
 
-program.hook("postAction", () => process.exit(0));
 program.parse();